add more broken unit tests

+++ tests/unit/src/com/nolanlawson/couchdroid/test/CrudTest.java
@@ -1,8 +1,10 @@
 package com.nolanlawson.couchdroid.test;
 
+import java.util.Arrays;
 import java.util.Random;
 
 import junit.framework.Assert;
+
 import android.annotation.SuppressLint;
 import android.test.ActivityInstrumentationTestCase2;
 import android.util.Log;
@@ -10,11 +12,15 @@ import android.util.Log;
 import com.nolanlawson.couchdroid.appforunittests.MainActivity;
 import com.nolanlawson.couchdroid.pouch.PouchDB;
 import com.nolanlawson.couchdroid.pouch.PouchException;
+import com.nolanlawson.couchdroid.pouch.PouchInfo;
 import com.nolanlawson.couchdroid.pouch.SynchronousPouchDB;
+import com.nolanlawson.couchdroid.test.data.GameBoy;
 import com.nolanlawson.couchdroid.test.data.Person;
 
 @SuppressLint("NewApi")
 public class CrudTest extends ActivityInstrumentationTestCase2<MainActivity>{
+    
+    private SynchronousPouchDB<Person> pouchDB;
     
     public CrudTest() {
         super(MainActivity.class);
@@ -27,6 +33,9 @@ public class CrudTest extends ActivityInstrumentationTestCase2<MainActivity>{
             Thread.sleep(100);
             Log.i("Tests", "Waiting for couch droid runtime to not be null");
         }
+        String dbName = "unit-test-" + Integer.toHexString(new Random().nextInt());
+        pouchDB = PouchDB.newSynchronousPouchDB(Person.class, 
+                getActivity().getCouchDroidRuntime(), dbName);
     }
 
 
@@ -34,43 +43,50 @@ public class CrudTest extends ActivityInstrumentationTestCase2<MainActivity>{
     @Override
     protected void tearDown() throws Exception {
         super.tearDown();
+        if (pouchDB != null) {
+            pouchDB.destroy();
+        }
     }
     
-    public void testCreateDestroy() throws PouchException {
-        String dbName = "unit-test-" + Integer.toHexString(new Random().nextInt());
+    public void testIdempotentPut() throws PouchException {
         
-        SynchronousPouchDB<Person> pouchDB = PouchDB.newSynchronousPouchDB(Person.class, 
-                getActivity().getCouchDroidRuntime(), dbName);
+        Person kenny = new Person("Kenny", 243987423, 4, null, false);
         
-        Person person = new Person("Mr. Mackey", 0, 0, null, false);
-        person.setPouchId("fooId");
-        pouchDB.put(person);
-        Person gotPerson = pouchDB.get("fooId");
-        assertEquals(person, gotPerson);
-        pouchDB.destroy();
         try {
-            pouchDB.get("fooId");
+            pouchDB.put(kenny);
             Assert.fail();
-        } catch (Exception expected) {
-            
+        } catch (PouchException expected) {
+            assertEquals(expected.getPouchError().getStatus(), 412); // missing_id, id required for puts
         }
+        kenny.setPouchId("kenny");
+        pouchDB.put(kenny);
         
-        // ok, re-create it and try to get the person
-        pouchDB = PouchDB.newSynchronousPouchDB(Person.class, 
-                getActivity().getCouchDroidRuntime(), dbName);
+        assertEquals(pouchDB.get("kenny"), kenny);
         
         try {
-            Person nonexistantPerson = pouchDB.get("fooId");
-            System.out.println(nonexistantPerson);
+            pouchDB.put(pouchDB.get("kenny"));
             Assert.fail();
         } catch (PouchException expected) {
             
         }
     }
+    
+    public void testPost() throws PouchException {
+        
+        Person kyle = new Person("Kyle", 24308, 2, null, true);
+        
+        PouchInfo pouchInfo = pouchDB.post(kyle);
+        
+        String kyleId = pouchInfo.getId();
+        
+        assertEquals(kyle, pouchDB.get(kyleId));
 
-    public void testPutPostGet() {
+        kyle = pouchDB.get(kyleId);
+        kyle.setGameBoys(Arrays.asList(new GameBoy("gba", "Game Boy Advance")));
+        kyle.setNumberOfPetsOwned(3);
         
+        pouchDB.post(kyle);
         
+        assertEquals(kyle, pouchDB.get(kyleId));
     }
-
 }
